Temporarily split ws and example transports to separate files
